+++ serial_commands.h
@@ -1,5 +1,8 @@
 // The parser - calls the appropriate function; leaves argument parsing to that function
 
+#ifndef SERIAL_COMMANDS_
+#define SERIAL_COMMANDS_
+
 typedef void cmdHandler_t();  // All handlers take nothing and return nothing
 
 struct cmd_table_t {
@@ -9,11 +12,14 @@ struct cmd_table_t {
 } ;
 
 // Define the command set
-extern cmdHandler_t cmd_batt, cmd_res, cmd_showcal, cmd_factor, cmd_offset, cmd_cal, cmd_activate, cmd_write, cmd_read, cmd_defaults, cmd_help;
+extern cmdHandler_t cmd_batt, cmd_res, cmd_adcref, cmd_adccal, cmd_showcal, cmd_factor, cmd_offset, cmd_cal, 
+                    cmd_activate, cmd_write, cmd_read, cmd_defaults, cmd_comment, cmd_help;
 
-const cmd_table_t cmd_table[] = {
+const cmd_table_t PROGMEM cmd_table[] = {
   {"batt", cmd_batt, "Show battery status."},
-  {"res", cmd_res, "Monitor ADC readings. Any input to stop."},
+  {"res", cmd_res, "<optional int> Monitor ADC readings - indicated number, or any input to stop."},
+  {"adcref", cmd_adcref, "Toggle the analog reference between default 3.6V and Vdd."},
+  {"adccal", cmd_adccal, "Calibrate the ADC offset."},
   {"showcal", cmd_showcal, "Show the calibration currently in use."},
   {"factor", cmd_factor, "<float> Enter a new cal factor (should not be necessary)."},
   {"offset", cmd_offset, "<float> Enter a new calibration offset (follow with activate)."},
@@ -22,6 +28,7 @@ const cmd_table_t cmd_table[] = {
   {"write", cmd_write, "Write currently active calibration to the calibration file. Requires confirmation."},
   {"read", cmd_read, "Read calibration from the calibration file."},
   {"defaults", cmd_defaults, "Set calibration to hard-coded defaults. Requires confirmation."},
+  {"c", cmd_comment, "Enter a comment (shows in the terminal log."},
   {"help", cmd_help, "This list."}
 } ;
 
@@ -55,3 +62,5 @@ void test()
     curr_cmd_set->cmd_table[1].cmdHandler();  // Call a member of the set in use
 }
 */
+
+#endif